# public/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginInputDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginOutputDTO"
                }
              }
            }
          }
        },
        "tags": [
          "Authorization"
        ]
      }
    },
    "/api/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterInputDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterOutputDTO"
                }
              }
            }
          }
        },
        "tags": [
          "Authorization"
        ]
      }
    },
    "/api/users": {
      "get": {
        "operationId": "UsersController_getUsers",
        "parameters": [
          {
            "name": "nameLike",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOutputDTO"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "AccessToken": []
          }
        ]
      }
    },
    "/api/users/me": {
      "get": {
        "operationId": "UsersController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOutputDTO"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "AccessToken": []
          }
        ]
      }
    },
    "/api/users/me/items": {
      "get": {
        "operationId": "UsersController_getMeItems",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/UserItemOutputDTO"
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "totalItems": {
                          "type": "number"
                        },
                        "itemCount": {
                          "type": "number"
                        },
                        "itemsPerPage": {
                          "type": "number"
                        },
                        "totalPages": {
                          "type": "number"
                        },
                        "currentPage": {
                          "type": "number"
                        }
                      },
                      "required": [
                        "totalItems",
                        "itemCount",
                        "itemsPerPage",
                        "totalPages",
                        "currentPage"
                      ]
                    }
                  },
                  "required": [
                    "items",
                    "meta"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "AccessToken": []
          }
        ]
      }
    },
    "/api/users/me/items/{id}/quick-sell": {
      "post": {
        "operationId": "UsersController_quickSellUserItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuickSoldUserItemOutputDTO"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "AccessToken": []
          }
        ]
      }
    },
    "/api/packs": {
      "get": {
        "operationId": "PacksController_getPacks",
        "parameters": [
          {
            "name": "id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "nameLike",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PackOutputDTO"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Packs"
        ],
        "security": [
          {
            "AccessToken": []
          }
        ]
      }
    },
    "/api/packs/{id}": {
      "get": {
        "operationId": "PacksController_getPack",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PackWithPokemonsOutputDTO"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "tags": [
          "Packs"
        ],
        "security": [
          {
            "AccessToken": []
          }
        ]
      }
    },
    "/api/packs/{id}/open": {
      "post": {
        "operationId": "PacksController_openPack",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenedPackOutputDTO"
                }
              }
            }
          },
          "404": {
            "description": ""
          },
          "409": {
            "description": ""
          }
        },
        "tags": [
          "Packs"
        ],
        "security": [
          {
            "AccessToken": []
          }
        ]
      }
    },
    "/api/trades": {
      "post": {
        "operationId": "TradesController_createPendingTrade",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePendingTradeInputDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingTradeOutputDTO"
                }
              }
            }
          }
        },
        "tags": [
          "Trades"
        ],
        "security": [
          {
            "AccessToken": []
          }
        ]
      }
    },
    "/api/trades/{id}/accept": {
      "post": {
        "operationId": "TradesController_acceptPendingTradeById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptedTradeOutputDTO"
                }
              }
            }
          }
        },
        "tags": [
          "Trades"
        ],
        "security": [
          {
            "AccessToken": []
          }
        ]
      }
    },
    "/api/trades/{id}/cancel": {
      "post": {
        "operationId": "TradesController_cancelPendingTradeById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancelledTradeOuputDTO"
                }
              }
            }
          }
        },
        "tags": [
          "Trades"
        ],
        "security": [
          {
            "AccessToken": []
          }
        ]
      }
    },
    "/api/trades/{id}/reject": {
      "post": {
        "operationId": "TradesController_rejectPendingTradeById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedTradeOutputDTO"
                }
              }
            }
          }
        },
        "tags": [
          "Trades"
        ],
        "security": [
          {
            "AccessToken": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "Poketrade API",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "AccessToken": {
        "type": "apiKey",
        "in": "header",
        "name": "x-access-token"
      }
    },
    "schemas": {
      "LoginInputDTO": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "LoginOutputDTO": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          }
        },
        "required": [
          "accessToken"
        ]
      },
      "RegisterInputDTO": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "confirmPassword": {
            "type": "string"
          }
        },
        "required": [
          "username",
          "password",
          "confirmPassword"
        ]
      },
      "RegisterOutputDTO": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          }
        },
        "required": [
          "accessToken"
        ]
      },
      "UserOutputDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "balance": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "name",
          "balance"
        ]
      },
      "PokemonOutputDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "worth": {
            "type": "number"
          },
          "height": {
            "type": "number"
          },
          "weight": {
            "type": "number"
          },
          "image": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "worth",
          "height",
          "weight",
          "image"
        ]
      },
      "UserItemOutputDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "receivedAt": {
            "format": "date-time",
            "type": "string"
          },
          "pokemon": {
            "$ref": "#/components/schemas/PokemonOutputDTO"
          }
        },
        "required": [
          "id",
          "receivedAt",
          "pokemon"
        ]
      },
      "QuickSoldUserItemOutputDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "receivedAt": {
            "format": "date-time",
            "type": "string"
          },
          "pokemon": {
            "$ref": "#/components/schemas/PokemonOutputDTO"
          },
          "soldAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "receivedAt",
          "pokemon",
          "soldAt"
        ]
      },
      "PackOutputDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "image": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "price",
          "image"
        ]
      },
      "PackWithPokemonsOutputDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "image": {
            "type": "string"
          },
          "pokemons": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PokemonOutputDTO"
            }
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "price",
          "image",
          "pokemons"
        ]
      },
      "OpenedPackOutputDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "openedAt": {
            "format": "date-time",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserOutputDTO"
          },
          "pack": {
            "$ref": "#/components/schemas/PackOutputDTO"
          },
          "pokemon": {
            "$ref": "#/components/schemas/PokemonOutputDTO"
          }
        },
        "required": [
          "id",
          "openedAt",
          "user",
          "pack",
          "pokemon"
        ]
      },
      "CreatePendingTradeInputDTO": {
        "type": "object",
        "properties": {
          "senderItemIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "receiverId": {
            "type": "string"
          },
          "receiverItemIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "senderItemIds",
          "receiverId",
          "receiverItemIds"
        ]
      },
      "PendingTradeOutputDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "sender": {
            "$ref": "#/components/schemas/UserOutputDTO"
          },
          "receiver": {
            "$ref": "#/components/schemas/UserOutputDTO"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "sender",
          "receiver",
          "status"
        ]
      },
      "AcceptedTradeOutputDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "sender": {
            "$ref": "#/components/schemas/UserOutputDTO"
          },
          "receiver": {
            "$ref": "#/components/schemas/UserOutputDTO"
          },
          "status": {
            "type": "string"
          },
          "acceptedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "sender",
          "receiver",
          "status",
          "acceptedAt"
        ]
      },
      "CancelledTradeOuputDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "sender": {
            "$ref": "#/components/schemas/UserOutputDTO"
          },
          "receiver": {
            "$ref": "#/components/schemas/UserOutputDTO"
          },
          "status": {
            "type": "string"
          },
          "cancelledAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "sender",
          "receiver",
          "status",
          "cancelledAt"
        ]
      },
      "RejectedTradeOutputDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "sender": {
            "$ref": "#/components/schemas/UserOutputDTO"
          },
          "receiver": {
            "$ref": "#/components/schemas/UserOutputDTO"
          },
          "status": {
            "type": "string"
          },
          "rejectedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "sender",
          "receiver",
          "status",
          "rejectedAt"
        ]
      }
    }
  }
}